// frontend/src/types/models/dashboards/Metrics.ts
export interface OrdersData {
  revenue: number;
  orders: number;
  graphData: {
    dataOne: {number: string; value: number}[];
    dataTwo: {number: string; value: number}[];
    dataThree: {number: string; value: number}[];
  };
}